move : 폴더 컨벤션에 따라 폴더명 변경
