+++ src/features/homepage/LogCard.tsx
