+++ src/pages/profile/index.tsx
@@ -1,6 +1,6 @@
 import PageContentLayout from '@/components/Layout/PageContentLayout';
-import ProfileHeader from '@/feature/profile/ProfileHeader';
-import TapNavigation from '@/feature/profile/TapNavigation';
+import ProfileHeader from '@/features/profile/ProfileHeader';
+import TapNavigation from '@/features/profile/TapNavigation';
 import { Outlet } from 'react-router-dom';
 
 function Profile() {
